backend/requirement.txt
# Flask Backend
flask>=3.0.0
flask-cors>=4.0.0

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Document Processing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# File Parsing
pypdf>=3.0.0
python-docx>=1.0.0

# Frontend
streamlit>=1.29.0

# Required for Windows compatibility
numpy>=1.24.0

# Required for Nomic Embeddings
einops
google-generativeai>=0.3.0
duckduckgo-search>=5.0.0
